⚡ Heuristic work
This one is great but does not guarantee to find the optimal count.
It works on my real input for part 2 but not for part 1.
However we move from 7.5s to 0.150s.

Note: changing the threshold from 4 (perfectly valid value) to 3 moves
from 7.5s to 3.5s, from 3 to 2: 3.5s -> 0.150s.

Change-Id: If979242f68e0670afd75850e1127914b933491eb